@@ -12,8 +12,8 @@
 		}
 	],
 	"require": {
-		"php": ">=5.4",
-		"arc/base": "~2.1"
+		"php": ">=7.1",
+		"arc/base": "~3.0"
 	},
 	"autoload": {
 		"psr-4": {
@@ -21,6 +21,9 @@
 		}
 	},
 	"require-dev": {
-		"phpunit/phpunit": "5.1.*"
+		"phpunit/phpunit": "9.*"
+	},
+	"scripts": {
+		"test": "phpunit --testdox"
 	}
 }
